+++ src/definitions/mod.rs
@@ -1,0 +1,43 @@
+use request::InvokeArgument;
+use response::InvokeResponse;
+use state::CucumberState;
+
+pub mod registration;
+
+pub trait SendableStep<World>: Send + Fn(&mut CucumberState, &mut World, Vec<InvokeArgument>) -> InvokeResponse {}
+
+impl<T, World> SendableStep<World> for T where T: Send + Fn(&mut CucumberState, &mut World, Vec<InvokeArgument>) -> InvokeResponse {}
+
+pub type Step<World> = Box<SendableStep<World, Output=InvokeResponse>>;
+pub type StepId = u32;
+
+// NOTE: These are capitalized to follow Cucumber general conventions, rather than Rust
+#[macro_export]
+macro_rules! Given {
+  ($cuke:ident; $regex:expr, $body:expr) => {{
+    use $crate::regex;
+    $cuke.given(file!(), line!(), regex::build($regex), Box::new(move |cuke, world, args| {
+      ($body)(cuke, world, try_destructure!(args))
+    }))
+  }}
+}
+
+#[macro_export]
+macro_rules! When {
+  ($cuke:ident; $regex:expr, $body:expr) => {{
+    use $crate::regex;
+    $cuke.when(file!(), line!(), regex::build($regex), Box::new(move |cuke, world, args| {
+      ($body)(cuke, world, try_destructure!(args))
+    }))
+  }}
+}
+
+#[macro_export]
+macro_rules! Then {
+  ($cuke:ident; $regex:expr, $body:expr) => {{
+    use $crate::regex;
+    $cuke.then(file!(), line!(), regex::build($regex), Box::new(move |cuke, world, args| {
+      ($body)(cuke, world, try_destructure!(args))
+    }))
+  }}
+}
