+++ src/destructuring/mod.rs
@@ -15,8 +15,8 @@ pub use self::invoke_arg_set::{
 #[macro_export]
 macro_rules! try_destructure {
   ($r: ident) => ({
-    use $crate::cucumber::InvokeResponse;
-    use $crate::cucumber::destructuring::{DestructurableSet};
+    use $crate::InvokeResponse;
+    use $crate::destructuring::{DestructurableSet};
 
     match $r.destructure_set() {
       Ok(e) => e,
