+++ src/lib.rs
@@ -6,12 +6,22 @@ extern crate hyper;
 extern crate serde;
 extern crate serde_json;
 
-pub mod cucumber;
-pub mod runner;
+mod request;
+mod response;
 mod server;
-pub mod regex;
 
-pub use cucumber::{ Step, Cucumber, CucumberRegistrar, InvokeResponse, InvokeArgument, FailMessage};
+pub mod definitions;
+pub mod destructuring;
+pub mod runner;
+pub mod regex;
+pub mod state;
+
+
+pub use definitions::Step;
+pub use definitions::registration::CucumberRegistrar;
+pub use request::InvokeArgument;
+pub use response::{InvokeResponse, FailMessage};
+pub use state::Cucumber;
 pub use runner::{ WorldRunner };
 pub use server::{ Server };
